@@ -1,10 +1,11 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 namespace Microsoft.Azure.Functions.Worker.Diagnostics
 {
     internal enum OpenTelemetrySchemaVersion
     {
-        v1_17_0
+        V1_17_0,
+        V1_37_0
     }
 }
