@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -88,7 +88,7 @@ namespace Microsoft.Azure.Functions.Worker.Logging
                 {
                     foreach (var pair in properties)
                     {
-                        if (pair.Key == FunctionInvocationScope.FunctionInvocationIdKey)
+                        if (pair.Key == TraceConstants.InternalKeys.FunctionInvocationId)
                         {
                             log.InvocationId = pair.Value?.ToString();
                             break;
