@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 namespace Microsoft.Azure.Functions.Worker.OpenTelemetry
@@ -12,5 +12,11 @@ namespace Microsoft.Azure.Functions.Worker.OpenTelemetry
         internal const string RegionNameEnvVar = "REGION_NAME";
         internal const string ResourceGroupEnvVar = "WEBSITE_RESOURCE_GROUP";
         internal const string OwnerNameEnvVar = "WEBSITE_OWNER_NAME";
+        internal const string WorkerDefaultSchemaVersion = "1.37.0";
+        internal const string WorkerActivitySourceName = "Microsoft.Azure.Functions.Worker";
+
+        // Capability variables
+        internal const string WorkerOTelEnabled = "WorkerOpenTelemetryEnabled";
+        internal const string WorkerOTelSchemaVersion = "WorkerOpenTelemetrySchemaVersion";
     }
 }
