@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System;
 using System.IO;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
@@ -10,27 +9,25 @@ using Xunit.Abstractions;
 
 namespace Microsoft.Azure.Functions.Sdk.E2ETests
 {
-    public sealed class InnerBuildTests(ITestOutputHelper testOutputHelper) : IDisposable
+    public class InnerBuildTests
     {
-        private readonly ProjectBuilder _builder = new(
-            testOutputHelper,
-            Path.Combine(TestUtility.TestResourcesProjectsRoot, "FunctionApp01", "FunctionApp01.csproj"));
+        private readonly ITestOutputHelper _testOutputHelper;
 
-        [Theory]
-        [InlineData("", false)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=true", false)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=true -p:FunctionsWriteMetadataJson=false", false)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=true -p:FunctionsWriteMetadataJson=true", true)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=false", true)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=false -p:FunctionsWriteMetadataJson=false", false)]
-        [InlineData("-p:FunctionsEnableWorkerIndexing=false -p:FunctionsWriteMetadataJson=true", true)]
-        public async Task Build_ScansReferences(string parameters, bool metadataGenerated)
+        public InnerBuildTests(ITestOutputHelper testOutputHelper)
         {
-            await _builder.RestoreAsync();
-            await _builder.BuildAsync(parameters, restore: false);
+            _testOutputHelper = testOutputHelper;
+        }
+
+        [Fact]
+        public async Task Build_ScansReferences()
+        {
+            string outputDir = await TestUtility.InitializeTestAsync(_testOutputHelper, nameof(Build_ScansReferences));
+            string projectFileDirectory = Path.Combine(TestUtility.TestResourcesProjectsRoot, "FunctionApp01", "FunctionApp01.csproj");
+
+            await TestUtility.RestoreAndBuildProjectAsync(projectFileDirectory, outputDir, null, _testOutputHelper);
 
             // Verify extensions.json contents
-            string extensionsJsonPath = Path.Combine(_builder.OutputPath, "extensions.json");
+            string extensionsJsonPath = Path.Combine(outputDir, "extensions.json");
             Assert.True(File.Exists(extensionsJsonPath));
 
             JToken extensionsJsonContents = JObject.Parse(File.ReadAllText(extensionsJsonPath));
@@ -57,13 +54,8 @@ namespace Microsoft.Azure.Functions.Sdk.E2ETests
             Assert.True(JToken.DeepEquals(expectedExtensionsJson, extensionsJsonContents));
 
             // Verify functions.metadata contents
-            string functionsMetadataPath = Path.Combine(_builder.OutputPath, "functions.metadata");
-            Assert.Equal(metadataGenerated, File.Exists(functionsMetadataPath));
-
-            if (!metadataGenerated)
-            {
-                return;
-            }
+            string functionsMetadataPath = Path.Combine(outputDir, "functions.metadata");
+            Assert.True(File.Exists(functionsMetadataPath));
 
             JToken functionsMetadataContents = JArray.Parse(File.ReadAllText(functionsMetadataPath));
             JToken expectedFunctionsMetadata = JArray.Parse(@"[
@@ -123,7 +115,5 @@ namespace Microsoft.Azure.Functions.Sdk.E2ETests
 
             Assert.True(JToken.DeepEquals(expectedFunctionsMetadata, functionsMetadataContents));
         }
-
-        public void Dispose() => _builder.Dispose();
     }
 }
